backend/requirements.txt
# SharePoint
Office365-REST-Python-Client==2.6.2

# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# HTTP
requests==2.33.1

# Auth — Microsoft/JWT
PyJWT==2.12.1
cryptography==47.0.0
msal==1.36.0

# Config e validação
pydantic==2.13.3
pydantic-settings==2.14.0
python-dotenv==1.2.2

# Testes
pytest==9.0.3
pytest-mock==3.15.1
